Move lat stiff and buckling code from ModeMatrix to theory
